+++ index.js
@@ -14,10 +14,4 @@ module.exports = {
     target.options.babel = target.options.babel || { includePolyfill: true };
     return this._super.included.apply(this, arguments);
   }
-
-  // contentFor(type, config) {
-  //   if (type === 'body-footer' && config.environment !== 'test') {
-  //     return '<div id="nucleus-modal-wormhole"></div>';
-  //   }
-  // }
 };
